backend/pyproject.toml
[project]
name = "pricecompare"
version = "0.1.0"
description = ""
authors = [
    {name = "ciire",email = "chueric9@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "cors (>=1.0.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "pymongo (>=4.13.0,<5.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "marshmallow (>=4.0.0,<5.0.0)"
]

packages = [
    { include = "backend" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
